# openapi.yaml
openapi: 3.0.2
info:
  title: OpenFairDB API
  version: 0.6.3
  contact:
    name: slowtec GmbH
    url: 'https://slowtec.de'
  license:
    name: AGPLv3
    url: 'https://github.com/slowtec/openfairdb/blob/master/LICENSE'
servers:
  - url: 'https://api.ofdb.io/v0/'
    description: Public production server
  - url: 'https://nightly.ofdb.io/v0/'
    description: Public unstable development server
paths:
...
  '/entries':
    post:
      summary: Create an entry
      tags:
        - Entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Entry'
      responses:
        '200':
          description: Successful response
  '/entries/{ids}':
    get:
      summary: Get multiple entries
      tags:
        - Entries
      parameters:
        - name: ids
          description: Comma-separated list of entry identifiers
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
...
  '/categories/{ids}':
    get:
      summary: Get multiple categories
      tags:
        - Categories
      parameters:
        - name: ids
          description: Comma-separated list of category identifiers
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
...
  '/login':
    post:
      summary: User login
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Sucessfull response
...
  '/users/reset-password-request':
    post:
      summary: Request a password reset
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
           description: Sucessfull response
  '/users/reset-password':
    post:
      summary: Request a users password
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                token:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
           description: Sucessfull response
...
  /export/entries.csv:
    get:
      summary: Export data as CSV
      description: |
        **Example**:

        Export all entries in Germany:
        `/export/entries.csv?bbox=47.49,0.79,54.63,18.30`

        If you want to find out the coordinates for other map areas,
        open "network" in the "developer tools" in your browser
        and look at the search request under at the value of `bbox`.
      tags:
        - Export
      parameters:
        - $ref: '#/components/parameters/BoundingBox'
      responses:
        '200':
          description: Successful response
          content:
            text/csv:
              schema:
                type: string
components:
  schemas:
    Entry:
      properties:
        id:
          type: string
          readOnly: true
        version:
          type: integer
        created:
          type: integer
        title:
          type: string
        description:
          type: string
        lat:
          type: number
        lng:
          type: number
        street:
          type: string
        zip:
          type: string
        city:
          type: string
        country:
          type: string
        email:
          type: string
        telephone:
          type: string
        homepage:
          type: string
        categories:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        ratings:
          type: array
          items:
            type: string
        image_url:
          type: string
        image_link_url:
          type: string
        license:
          type: string
    Category:
      properties:
        id:
          type: string
          readOnly: true
        created:
          type: integer
        version:
          type: integer
        name:
          type: string
...
  parameters:
    BoundingBox:
      name: bbox
      in: query
      description: Bounding Box
      schema:
        type: string
        example: '42.27,-7.97,52.58,38.25'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...